+++ src/host-relative-text-insertion.ts
@@ -1,6 +1,8 @@
-import { default as insertExcelText } from "./taskpane/excel-office-document";
-import { default as insertPowerPointText } from "./taskpane/powerpoint-office-document";
-import { default as insertWordText } from "./taskpane/word-office-document";
+// import { default as insertExcelText } from "./taskpane/excel-office-document";
+// import { default as insertPowerPointText } from "./taskpane/powerpoint-office-document";
+// import { default as insertWordText } from "./taskpane/word-office-document";
+// import { default as insertOutlookText } from "./taskpane/outlook-office-document";
+import { insertWordText, insertExcelText, insertPowerPointText, insertOutlookText } from "./taskpane/office-document";
 
 /* global Office */
 
@@ -20,6 +22,9 @@ export const selectInsertionByHost = async () => {
       case Office.HostType.Word: {
         insertText = insertWordText;
         break;
+      }
+      case Office.HostTyope.Outlook: {
+        insertText = insertOutlookText;
       }
       default: {
         throw new Error("There is no end-to-end test for that host.");
