--- src/taskpane/outlook-office-document.ts
@@ -1,20 +1,0 @@
-/* global Office console */
-
-const insertText = async (text: string) => {
-  // Write text to the cursor point in the compose surface.
-  try {
-    Office.context.mailbox.item.body.setSelectedDataAsync(
-      text,
-      { coercionType: Office.CoercionType.Text },
-      (asyncResult: Office.AsyncResult<void>) => {
-        if (asyncResult.status === Office.AsyncResultStatus.Failed) {
-          throw asyncResult.error.message;
-        }
-      }
-    );
-  } catch (error) {
-    console.log("Error: " + error);
-  }
-};
-
-export default insertText;
