--- src/taskpane/powerpoint-office-document.ts
@@ -1,16 +1,0 @@
-/* global Office console */
-
-const insertText = async (text: string) => {
-  // Write text to the selected slide.
-  try {
-    Office.context.document.setSelectedDataAsync(text, (asyncResult: Office.AsyncResult<void>) => {
-      if (asyncResult.status === Office.AsyncResultStatus.Failed) {
-        throw asyncResult.error.message;
-      }
-    });
-  } catch (error) {
-    console.log("Error: " + error);
-  }
-};
-
-export default insertText;
